FE/Qt4-OSX: seems to be necessary
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@31235 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/darwin/VBoxUtils-darwin.h
@@ -83,7 +83,7 @@ bool darwinIsApplicationCommand(ConstNativeNSEventRef pEvent);
 RT_C_DECLS_END
 
 DECLINLINE(CGRect) darwinToCGRect(const QRect& aRect) { return CGRectMake(aRect.x(), aRect.y(), aRect.width(), aRect.height()); }
-DECLINLINE(CGRect) darwinFlipCGRect(CGRect aRect, int aTargetHeight) { aRect.origin.y = aTargetHeight - aRect.origin.y - aRect.size.height; return aRect; }
+DECLINLINE(CGRect) darwinFlipCGRect(CGRect aRect, double aTargetHeight) { aRect.origin.y = aTargetHeight - aRect.origin.y - aRect.size.height; return aRect; }
 DECLINLINE(CGRect) darwinFlipCGRect(CGRect aRect, const CGRect &aTarget) { return darwinFlipCGRect(aRect, aTarget.size.height); }
 DECLINLINE(CGRect) darwinCenterRectTo(CGRect aRect, const CGRect& aToRect)
 {
